pull S3Manager logic out of SubmissionsExport and put in new S3Manager::Resource module, refactor some filename building logic in SubmissionsExport, move some filename logic out of performer
