+++ app/models/submissions_export.rb
@@ -25,17 +25,19 @@ class SubmissionsExport < ActiveRecord::Base
   validates :course_id, presence: true
   validates :assignment_id, presence: true
 
-  before_create :set_s3_bucket_name
+  before_save :rebuild_s3_object_key, if: :export_filename_changed?
 
-  def build_s3_object_key(object_filename)
-    if Rails.env.development?
-      [ ENV["AWS_S3_DEVELOPER_TAG"] ].concat(s3_object_key_pieces(object_filename)).join("/")
-    else
-      s3_object_key_pieces(object_filename).join("/")
-    end
+  def rebuild_s3_object_key
+    self[:s3_object_key] = build_s3_object_key(export_filename)
   end
 
-  def s3_object_key_pieces(object_filename)
+  def build_s3_object_key(object_filename)
+    key_pieces = [ s3_object_key_prefix, object_filename ].flatten
+    key_pieces.unshift ENV["AWS_S3_DEVELOPER_TAG"] if Rails.env.development?
+    key_pieces.join "/"
+  end
+
+  def s3_object_key_prefix
     [
       "exports",
       "courses",
@@ -43,15 +45,8 @@ class SubmissionsExport < ActiveRecord::Base
       "assignments",
       assignment_id,
       created_at_date,
-      created_at_in_microseconds,
-      object_filename
+      created_at_in_microseconds
     ]
-  end
-
-  def presigned_s3_url
-    if s3_object_key
-      s3_manager.bucket.object(s3_object_key).presigned_url(:get, expires_in: 604800).to_s
-    end
   end
 
   def created_at_in_microseconds
@@ -62,55 +57,7 @@ class SubmissionsExport < ActiveRecord::Base
     created_at.strftime("%F")
   end
 
-  def s3_manager
-    @s3_manager ||= S3Manager::Manager.new
-  end
-
   def downloadable?
     !!last_export_completed_at
-  end
-
-  def upload_file_to_s3(file_path)
-    s3_manager.put_encrypted_object(s3_object_key, file_path)
-  end
-
-  def fetch_object_from_s3
-    s3_manager.get_encrypted_object(s3_object_key)
-  end
-
-  def stream_s3_object_body
-    s3_object = fetch_object_from_s3
-    return unless s3_object and s3_object.body
-    s3_object.body.read
-  end
-
-  def write_s3_object_to_file(target_file_path)
-    s3_manager.write_encrypted_object_to_file(s3_object_key, target_file_path)
-  end
-
-  def delete_object_from_s3
-    s3_manager.delete_object(s3_object_key)
-  end
-
-  def update_export_completed_time
-    update_attributes last_export_completed_at: export_time
-  end
-
-  def set_s3_bucket_name
-    self[:s3_bucket_name] = s3_manager.bucket_name
-  end
-
-  def s3_object_exists?
-    s3_object_summary.exists?
-  end
-
-  def s3_object_summary
-    @s3_object_summary ||= S3Manager::Manager::ObjectSummary.new(s3_object_key, s3_manager)
-  end
-
-  private
-
-  def export_time
-    Time.now
   end
 end
